Refactored names in the server side.
