+++ Locationapp/src/java/locationrestfulservice/UserPK.java
@@ -3,7 +3,7 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package ticketerrestfulservice;
+package locationrestfulservice;
 
 import java.io.Serializable;
